backend/requirements.txt
# Backend Dependencies for AI Tutor (100% FREE - No OpenAI)
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Google Gemini (FREE API - No credit card needed!)
google-generativeai==0.8.3

# LangChain Core (minimal)
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.15
langchain-google-genai==2.0.5
langchain-text-splitters==0.3.2

# Vector Database (CPU only, lightweight)
faiss-cpu==1.12.0

# Document Processing
pypdf==5.1.0

# Speech Recognition (FREE - Uses Google)
SpeechRecognition==3.12.0

# Text-to-Speech (FREE - pyttsx3)
pyttsx3==2.98

# Environment Variables
python-dotenv==1.0.1

# Additional dependencies
pydantic==2.9.2
numpy==2.1.3
requests==2.32.3